refactor: share live device coverage helpers
