+++ .github/workflows/replays-nightly.yml
@@ -134,13 +134,10 @@ jobs:
             test -n "$ANDROID_SERIAL"
             BOOT_FINISHED_AT="$(date +%s)"
             pnpm build
-            pnpm build:android
             pnpm clean:daemon
             FULL_STARTED_AT="$(date +%s)"
             AGENT_DEVICE_ANDROID_E2E=1 AGENT_DEVICE_ANDROID_E2E_TIER=full AGENT_DEVICE_ANDROID_SERIAL="$ANDROID_SERIAL" AGENT_DEVICE_FIXTURE_APP_PATH="${{ steps.fixture-app.outputs.apk-path }}" AGENT_DEVICE_FIXTURE_APP_ID="${{ steps.fixture-app.outputs.app-id }}" node --test test/integration/smoke-android-emulator.test.ts
             FULL_FINISHED_AT="$(date +%s)"
-            COVERAGE_REPORT="$(find test/artifacts/android-emulator -name coverage-report.json -print | tail -n 1)"
-            node --input-type=module -e 'import fs from "node:fs"; const report = JSON.parse(fs.readFileSync(process.argv[1], "utf8")); console.log(`Android full scenario body: ${report.timings.runDurationMs}ms`); for (const scenario of report.timings.scenarios) console.log(`  ${scenario.id}: ${scenario.durationMs}ms`);' "$COVERAGE_REPORT"
             echo "Android setup before emulator: $((BOOT_FINISHED_AT - ${{ steps.android-setup.outputs.seconds }}))s"
             echo "Android full scenario wall time: $((FULL_FINISHED_AT - FULL_STARTED_AT))s"
             node --experimental-strip-types src/bin.ts test test/integration/replays/android/01-settings.ad --retries 2 --artifacts-dir test/artifacts/replays-android-settings --report-junit test/artifacts/replays-android-settings.junit.xml
