+++ test/integration/smoke-ios-simulator-coverage.test.ts
@@ -18,7 +18,7 @@ import {
   IOS_SIMULATOR_E2E_COVERAGE,
   liveCommandsForScenario,
 } from './ios-simulator-e2e/coverage-manifest.ts';
-import { collectPagedEventTimeline } from './ios-simulator-e2e/event-timeline.ts';
+import { collectPagedEventTimeline } from './live-device-e2e/event-timeline.ts';
 import { findMissingFixtureIdentifiers } from './ios-simulator-e2e/fixture-identifier-coverage.ts';
 import { IOS_SIMULATOR_LIVE_SCENARIOS } from './ios-simulator-e2e/scenarios.ts';
